requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python
numpy
pillow
requests
python-dotenv
scikit-learn
joblib
